Fixed script input
Running the script in the current state gives this error from perl:
```
readline() on unopened filehandle IN at sample_qfast_file.pl line 42.
````
This is because the ```open``` statement in the original script has been deleted which means IN is not a valid filehandle. The correct handle for reading from stdin is STDIN, which I have added in this commit.